+++ vbscan.pl
@@ -1,9 +1,9 @@
 #!/usr/bin/perl
 #
 #            --------------------------------------------------
-#                            OWASP VBScan
+#                              OWASP VBScan
 #            --------------------------------------------------
-#        Copyright (C) <2015>  <(Mohammad Reza Espargham)>
+#        Copyright (C) <2018>  <(Mohammad Reza Espargham)>
 #
 #        This program is free software: you can redistribute it and/or modify
 #        it under the terms of the GNU General Public License as published by
@@ -19,57 +19,55 @@
 #        along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 #
-# About Project Leader :
-#
-#    Project Leader     :   Mohammad Reza Espargham
-#    Linkedin    :   https://ir.linkedin.com/in/rezasp
-#    E-Mail      :   me@reza.es , reza.espargham@gmail.com , reza.espargham@owasp.org
-#    Twitter     :   https://twitter.com/rezesp
-#    SourceForge :   https://sourceforge.net/projects/vbscan/
-#    Github      :   https://github.com/rezasp/vbscan/
 
 $author="Mohammad Reza Espargham";$author.="";
-$version="0.1.7.1";$version.="";
-$codename="Larry Wall";$codename.=""; # https://en.wikipedia.org/wiki/Larry_Wall <3
-$update="2016/10/21";$update.="";
+$version="0.1.8";$version.="";
+$codename="Self Challenge";$codename.="";
+$update="2018/09/13";$update.="";
 
 system(($^O eq 'MSWin32') ? 'cls' : 'clear');
 use if $^O eq "MSWin32", Win32::Console::ANSI;
 use Term::ANSIColor;
-
+use Cwd;                                                                       
+$mepath = Cwd::realpath($0); $mepath =~ s#/[^/\\]*$##; 
 
 $SIG{INT} = \&interrupt;
 sub interrupt {
     fprint("\nShutting Down , Interrupt by user");
-    do "./core/report.pl";
+    do "$mepath/core/report.pl";
     print color("reset");
     exit 0;
 }
-do "./core/header.pl";
-do "./core/main.pl";
-do "./modules/firewall.pl";
-do "./core/ver.pl";
-do "./exploit/verexploit.pl";
-do "./modules/license.pl";
-do "./modules/pathdisclure.pl";
-do "./modules/dirlisting.pl";
-do "./modules/missconfig.pl";
-do "./modules/cpfinder.pl";
-do "./modules/cpupgrade.pl";
-do "./modules/validator.pl";
-do "./modules/robots.pl";
-do "./modules/c99finder.pl";
-do "./modules/backupfinder.pl";
-do "./modules/errfinder.pl";
-do "./modules/configfinder.pl";
-do "./exploit/faq.pl";
-do "./exploit/vbseo.pl";
-do "./exploit/xperience.pl";
-do "./exploit/upexploit.pl";
-do "./exploit/arcade.pl";
-do "./exploit/yui.pl";
-do "./exploit/htmlcode.pl";
-do "./exploit/decodearguments.pl";
+do "$mepath/core/header.pl";
+do "$mepath/core/main.pl";
+do "$mepath/modules/firewall.pl";
+do "$mepath/core/ver.pl";
+do "$mepath/exploit/verexploit.pl";
+do "$mepath/modules/passlogger.pl";
+do "$mepath/modules/license.pl";
+do "$mepath/modules/pathdisclure.pl";
+do "$mepath/modules/dirlisting.pl";
+do "$mepath/modules/missconfig.pl";
+do "$mepath/modules/cpfinder.pl";
+do "$mepath/modules/cpupgrade.pl";
+do "$mepath/modules/validator.pl";
+do "$mepath/modules/robots.pl";
+do "$mepath/modules/c99finder.pl";
+do "$mepath/modules/backupfinder.pl";
+do "$mepath/modules/errfinder.pl";
+do "$mepath/modules/configfinder.pl";
+do "$mepath/exploit/faq.pl";
+do "$mepath/exploit/vbseo.pl";
+do "$mepath/exploit/xperience.pl";
+do "$mepath/exploit/upexploit.pl";
+do "$mepath/exploit/arcade.pl";
+do "$mepath/exploit/regplug.pl";
+do "$mepath/exploit/tapatalk.pl";
+do "$mepath/exploit/routestring.pl";
+do "$mepath/exploit/openredirection.pl";
+do "$mepath/exploit/yui.pl";
+do "$mepath/exploit/htmlcode.pl";
+do "$mepath/exploit/decodearguments.pl";
 
-do "./core/report.pl";
+do "$mepath/core/report.pl";
 print color("reset");